+++ packages/apollo-server-testing/src/createTestClient.ts
@@ -1,6 +1,8 @@
 import { ApolloServerBase } from 'apollo-server-core';
 import { print, DocumentNode } from 'graphql';
 
+type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>;
+
 type StringOrAst = string | DocumentNode;
 
 // A query must not come with a mutation (and vice versa).
@@ -9,7 +11,12 @@ type Mutation = { mutation: StringOrAst; query?: undefined };
 
 export default (server: ApolloServerBase) => {
   const executeOperation = server.executeOperation.bind(server);
-  const test = ({ query, mutation, ...args }: Query | Mutation) => {
+  const test = ({
+    query,
+    mutation,
+    ...args
+  }: (Query | Mutation) &
+    Omit<Parameters<typeof server.executeOperation>[0], 'query'>) => {
     const operation = query || mutation;
 
     if (!operation || (query && mutation)) {