+++ packages/apollo-server-testing/CHANGELOG.md
@@ -2,4 +2,6 @@
 
 ### vNEXT
 
+* Add variables prop to Query and Mutation types. Fixes #2172
+
 * `apollo-server-testing`: Added createTestClient function